+++ includes/plugin/class-core.php
@@ -113,6 +113,7 @@ class Core {
 		$this->loader->add_action( 'wp_ajax_hide_decalog_nag', $nag, 'hide_callback' );
 		$this->loader->add_filter( 'myblogs_blog_actions', $plugin_admin, 'blog_action', 10, 2 );
 		$this->loader->add_filter( 'manage_sites_action_links', $plugin_admin, 'site_action', 10, 3 );
+		add_shortcode( 'decalog-selfreg', [ 'Decalog\Plugin\Feature\SDK', 'sc_get_selfreg' ] );
 	}
 
 	/**